requirements.txt
twikit>=2.0.0
anthropic>=0.40.0
jinja2>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
